refactor: extract rehearsal song instrument command usecase
