+++ bandlog-rehearsal/src/main/kotlin/net/effize/bandlog/rehearsal/service/RehearsalCommandUseCase.kt
@@ -1,16 +1,10 @@
 package net.effize.bandlog.rehearsal.service
 
 import net.effize.bandlog.rehearsal.adapter.out.team.TeamAdapter
-import net.effize.bandlog.rehearsal.dto.request.AddRehearsalSongInstrumentRequest
-import net.effize.bandlog.rehearsal.dto.request.AssignMemberToInstrumentRequest
 import net.effize.bandlog.rehearsal.dto.request.CreateRehearsalRequest
 import net.effize.bandlog.rehearsal.dto.request.ModifyRehearsalRequest
-import net.effize.bandlog.rehearsal.dto.request.ModifyRehearsalSongInstrumentRequest
-import net.effize.bandlog.rehearsal.dto.response.AddRehearsalSongInstrumentResponse
-import net.effize.bandlog.rehearsal.dto.response.AssignMemberToInstrumentResponse
 import net.effize.bandlog.rehearsal.dto.response.CreateRehearsalResponse
 import net.effize.bandlog.rehearsal.dto.response.ModifyRehearsalResponse
-import net.effize.bandlog.rehearsal.dto.response.ModifyRehearsalSongResponse
 import net.effize.bandlog.rehearsal.model.Rehearsal
 import net.effize.bandlog.rehearsal.repository.RehearsalRepository
 import net.effize.bandlog.shared.auth.AuthUser
@@ -67,73 +61,6 @@ class RehearsalCommandUseCase(
 
 
     @Transactional
-    fun addRehearsalSongInstrument(
-        authUser: AuthUser,
-        rehearsalId: Long,
-        rehearsalSongId: Long,
-        addRehearsalSongInstrumentRequest: AddRehearsalSongInstrumentRequest
-    ): AddRehearsalSongInstrumentResponse {
-        if (!teamAdapter.isUserLeaderOfTeam(authUser.id, addRehearsalSongInstrumentRequest.teamId)) {
-            throw IllegalStateException("Cannot add instrument to rehearsal you are not leader of")
-        }
-
-        val rehearsal = rehearsalRepository.findByIdOrNull(rehearsalId)
-            ?: throw IllegalArgumentException("Rehearsal with id $rehearsalId not found")
-
-        rehearsal.addInstrumentToSong(rehearsalSongId, addRehearsalSongInstrumentRequest.instrument)
-
-        return AddRehearsalSongInstrumentResponse(rehearsal.id)
-    }
-
-    @Transactional
-    fun modifyRehearsalSongInstrument(
-        authUser: AuthUser,
-        rehearsalId: Long,
-        rehearsalSongId: Long,
-        rehearsalSongInstrumentId: Long,
-        modifyRehearsalSongInstrumentRequest: ModifyRehearsalSongInstrumentRequest
-    ): ModifyRehearsalSongResponse {
-        if (!teamAdapter.isUserLeaderOfTeam(authUser.id, modifyRehearsalSongInstrumentRequest.teamId)) {
-            throw IllegalStateException("Cannot add instrument to rehearsal you are not leader of")
-        }
-
-        val rehearsal = rehearsalRepository.findByIdOrNull(rehearsalId)
-            ?: throw IllegalArgumentException("Rehearsal with id $rehearsalId not found")
-
-        rehearsal.modifySongInstrument(
-            rehearsalSongId,
-            rehearsalSongInstrumentId,
-            modifyRehearsalSongInstrumentRequest.instrument,
-            modifyRehearsalSongInstrumentRequest.order
-        )
-
-        return ModifyRehearsalSongResponse(rehearsal.id)
-    }
-
-    @Transactional
-    fun assignMember(
-        authUser: AuthUser,
-        rehearsalId: Long,
-        rehearsalSongId: Long,
-        rehearsalSongInstrumentId: Long,
-        assignMemberRequest: AssignMemberToInstrumentRequest
-    ): AssignMemberToInstrumentResponse {
-        if (!teamAdapter.isUserLeaderOfTeam(authUser.id, assignMemberRequest.teamId)) {
-            throw IllegalStateException("Cannot assign member to rehearsal you are not leader of")
-        }
-        if (!teamAdapter.isMemberOfTeam(assignMemberRequest.memberId, assignMemberRequest.teamId)) {
-            throw IllegalStateException("Cannot assign member to rehearsal you are not a member of")
-        }
-
-        val rehearsal = rehearsalRepository.findByIdOrNull(rehearsalId)
-            ?: throw IllegalArgumentException("Rehearsal with id $rehearsalId not found")
-
-        rehearsal.assignMemberToInstrument(rehearsalSongId, rehearsalSongInstrumentId, assignMemberRequest.memberId)
-
-        return AssignMemberToInstrumentResponse(rehearsal.id)
-    }
-
-    @Transactional
     fun deleteRehearsal(authUser: AuthUser, rehearsalId: Long) {
         val rehearsal = rehearsalRepository.findByIdOrNull(rehearsalId)
             ?: throw IllegalArgumentException("Rehearsal with id $rehearsalId not found")
@@ -143,23 +70,5 @@ class RehearsalCommandUseCase(
         }
 
         rehearsalRepository.deleteById(rehearsalId)
-    }
-
-
-    @Transactional
-    fun deleteRehearsalSongInstrument(
-        authUser: AuthUser,
-        rehearsalId: Long,
-        rehearsalSongId: Long,
-        rehearsalSongInstrumentId: Long
-    ) {
-        val rehearsal = rehearsalRepository.findByIdOrNull(rehearsalId)
-            ?: throw IllegalArgumentException("Rehearsal with id $rehearsalId not found")
-
-        if (!teamAdapter.isUserLeaderOfTeam(authUser.id, rehearsal.teamId)) {
-            throw IllegalStateException("Cannot delete rehearsal song instrument you are not leader of")
-        }
-
-        rehearsal.deleteSongInstrument(rehearsalSongId, rehearsalSongInstrumentId)
     }
 }
